Move asset table to new file
